libffi: C64 initial commit
libffi version 3.4.2 from CHERIBSD repo:
https://github.com/CTSRD-CHERI/libffi/commit/c666fdd8613b4d56130431c3e8f0bb05f30f07e8

Included patch to add compile time option `--enable-cheri-ffi-closures`
to enable ffi closures; this functionality was disabled in the source
repo as it was not fully implemented, but needs to be enabled for
Glib/GDBus signals to work. Use of FFI closures outside of GDBus signals
has not been tested so use at your own risk.

Signed-off-by: Michael Cobb <michael.cobb@iceotope.com>